+++ src/api/index.ts
@@ -1,32 +1,43 @@
-import axios from 'axios';
+import axios, { AxiosError } from 'axios';
 
 import { Rank } from '@/types/rank';
 
-const sheetId =
-    'AKfycbzIxhaykKSHsUxzgTFfsurMuv6o2oAtAqNkBqMnCkwbbWk6KXfKjvV8ZBCRAqjVv3j6cQ';
+const baseUrl = `/api`;
 
-const googleAppScriptUrl = `https://script.google.com/macros/s/${sheetId}/exec`;
-const sheetName = 'records_2024';
-const baseUrl = `${googleAppScriptUrl}?Sheet=${sheetName}`;
+const errorHandler = async (error: AxiosError) => {
+    if (error.isAxiosError) {
+        console.error('axios error: \n', error);
+    }
+
+    return Promise.reject(error);
+};
+
+axios.interceptors.response.use((response) => response, errorHandler);
 
 /** 取得 sheet 資料 */
-const getData = async (): Promise<string[][]> => {
-    const url = [baseUrl, 'Operation=get'].join('&');
-
-    return axios.get(url).then((res) => res.data);
+const getData = async (): Promise<{ data: string[][] }> => {
+    return await axios({
+        baseURL: baseUrl,
+        method: 'get',
+        timeout: 60000,
+    });
 };
 
 /** 更新 sheet 資料 */
-const updateData = async (data: Rank): Promise<any> => {
-    const url = [
-        baseUrl,
-        'Operation=update',
-        `playerName=${data.playerName}`,
-        `player=${JSON.stringify(data.player)}`,
-        `lastWords=${data.lastWords}`,
-    ].join('&');
+const updateData = async (data: Rank): Promise<{ data: string[][] }> => {
+    const mutatedData: { [key: string]: string } = {};
 
-    return axios.get(url).then((res) => res);
+    Object.keys(data).forEach((key) => {
+        // @ts-ignore
+        mutatedData[key] = JSON.stringify(data[key]);
+    });
+
+    return await axios({
+        baseURL: baseUrl,
+        method: 'post',
+        timeout: 60000,
+        data: mutatedData,
+    });
 };
 
 export default {