+++ vite.config.ts
@@ -1,7 +1,4 @@
-import {
-    fileURLToPath,
-    URL,
-} from 'node:url';
+import { fileURLToPath, URL } from 'node:url';
 
 import { defineConfig } from 'vite';
 import vuetify, { transformAssetUrls } from 'vite-plugin-vuetify';
@@ -23,8 +20,20 @@ export default defineConfig({
     define: { 'process.env': {} },
     resolve: {
         alias: [
-            { find: '@', replacement: fileURLToPath(new URL('./src', import.meta.url)) },
+            {
+                find: '@',
+                replacement: fileURLToPath(new URL('./src', import.meta.url)),
+            },
         ],
         extensions: ['.js', '.json', '.jsx', '.mjs', '.ts', '.tsx', '.vue'],
+    },
+    server: {
+        proxy: {
+            '/api': {
+                target: 'https://script.google.com/macros/s/AKfycbxxYhVKAzrI7aYZ4PSC7_8q9QeXSkmmicVXRaJV5JjCz43C8DhjlMsL5c2kCaLNSjCguw/exec',
+                changeOrigin: true,
+                rewrite: (path) => path.replace(/^\/api/, ''),
+            },
+        },
     },
 });